moved mc_facet_geometry to new module
