+++ src/geometry/mod.rs
@@ -9,7 +9,6 @@ pub mod grid_assignment_object;
 pub mod mc_cell_state;
 pub mod mc_domain;
 pub mod mc_facet_adjacency;
-pub mod mc_facet_geometry;
 pub mod mc_location;
 pub mod mc_nearest_facet;
 pub mod mesh_partition;
